Initial code import
KaeruErr is an error aggregator, similar to Errbit, but designed
to be simpler.  Kaeru is japanese for frog/toad and Err is japanese
for error.

This code was all written in 3 and a half hours during the SacRuby
May 2013 hackathon.  There are quite a few bugs and no tests, but
it already has the following features:

* Add error information directly to the database.
* Upload error information to external site via HTTP.
* Sinatra application for receiving error information and viewing
  errors.
* Full-text search through all related error fields.

I plan on fixing the bugs and adding tests, and well as adding
some more features, but the goal is definitely to keep things
simple.