More discrete searching features
Instead of shoving everything in a text column and doing a full text
search on that, use separate indexes for each column for specific
searching.

Use forme to build the forms on the search page.

Make backtrace lines clickable on error page to do a search for
other errors with the same backtrace line.

Show status on error pages, so you can see if the errors are opened
or closed.

Add a notes column, which will be used to add notes about errors.