Add specs and Rakefile with spec and migration tasks
There are four different spec suites.  One is for the recorder
and one for the reporter.  The third is model spec tests for
the web application, and the last are tests of the web
application itself.  The recorder and model specs use plain
rspec and Sequel.  The reporter spec tests against a web
server running the web application with unicorn.  The web
specs test with capybara.  The specs are all written for
rspec 1, though migrating to rspec 2 should be easy if that
is ever desired.

The Rakfile has tasks to run the specs, with the default task
running all four spec suites.  It has migration tasks for
migrating the databases up and down in each environment,
except for production, where migrating down is not purposely
not supported.