Use Roda typecast_params plugin for all params access
This is simpler and less likely to have security vulnerabilities.
Specifically, see the changes when searching for errors, where
typecasting is done in the controller and the model can assume
correct types.

Remove use of indifferent params, as they make it harder to
separate what is considered external from what is considered
internal.