Use SimpleCov.add_filter block instead of string
Using a string is almost always a bad idea, because if any part of
the path contains the string, it will filter out the path. Switch
to using a block and doing a match on the entire path prefix.

Additionally, instead of just filtering out spec files, filter out
all files except those in the lib directory, since this is only
used to check coverage for the files in the kaeruera gem.